Accumulate filter-bin stride product in int64_t in set_strides()
set_strides() multiplied the per-filter bin counts into a plain int accumulator, so the running product overflowed before being stored into the now-int64_t n_filter_bins_. Promote the accumulator to int64_t so the product is formed in 64-bit.

+++ src/tallies/tally.cpp
@@ -512,7 +512,7 @@ void Tally::set_strides()
   // longest stride.
   auto n = filters_.size();
   strides_.resize(n, 0);
-  int stride = 1;
+  int64_t stride = 1;
   for (int i = n - 1; i >= 0; --i) {
     strides_[i] = stride;
     stride *= model::tally_filters[filters_[i]]->n_bins();
